refactor: 테이블/테이블 그룹 의존성 제거 중간 커밋
