+++ src/main/java/kitchenpos/table/domain/OrderTables.java
@@ -2,13 +2,17 @@ package kitchenpos.table.domain;
 
 import kitchenpos.table.dto.OrderTableResponse;
 
+import javax.persistence.Embeddable;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
 import java.util.stream.Collectors;
 
+@Embeddable
 public class OrderTables {
     private List<OrderTable> value = new ArrayList<>();
+
+    protected OrderTables() {}
 
     public OrderTables(Collection<OrderTable> orderTables) {
         this.value.addAll(orderTables);
@@ -17,6 +21,17 @@ public class OrderTables {
     public List<OrderTableResponse> toResponse() {
         return this.value.stream()
                 .map(OrderTableResponse::from)
+                .collect(Collectors.toList());
+    }
+
+    public void ungroup() {
+        this.value.forEach(OrderTable::leaveGroup);
+    }
+
+    public List<Long> orderTableIds() {
+        return this.value.stream()
+                .mapToLong(OrderTable::getId)
+                .boxed()
                 .collect(Collectors.toList());
     }
