+++ src/main/java/kitchenpos/tableGroup/domain/TableGroup.java
@@ -1,0 +1,52 @@
+package kitchenpos.tableGroup.domain;
+
+import kitchenpos.table.domain.OrderTable;
+import kitchenpos.table.domain.OrderTables;
+
+import javax.persistence.*;
+import java.time.LocalDateTime;
+import java.util.List;
+
+@Entity
+public class TableGroup {
+
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    private LocalDateTime createdDate;
+
+    @Embedded
+    private OrderTables orderTables;
+
+    protected TableGroup() {}
+
+    public TableGroup(Long id, LocalDateTime createdDate, List<OrderTable> orderTables) {
+        this.id = id;
+        this.createdDate = createdDate;
+        this.orderTables = new OrderTables(orderTables);
+
+        orderTables.forEach(orderTable -> orderTable.joinGroup(this));
+    }
+
+    public TableGroup(OrderTables orderTables) {
+        this(null, LocalDateTime.now(), orderTables.getValue());
+    }
+
+    public void ungroup() {
+        this.orderTables.ungroup();
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public LocalDateTime getCreatedDate() {
+        return createdDate;
+    }
+
+    public OrderTables getOrderTables() {
+        return orderTables;
+    }
+}
