+++ src/test/java/kitchenpos/Acceptance/tableGroup/TableGroupAcceptanceTest.java
@@ -8,8 +8,8 @@ import kitchenpos.order.dto.OrderCreateRequest;
 import kitchenpos.order.dto.OrderLineItemRequest;
 import kitchenpos.order.dto.OrderResponse;
 import kitchenpos.table.dto.OrderTableResponse;
-import kitchenpos.table.dto.TableGroupCreateRequest;
-import kitchenpos.table.dto.TableGroupResponse;
+import kitchenpos.tableGroup.dto.TableGroupCreateRequest;
+import kitchenpos.tableGroup.dto.TableGroupResponse;
 import org.apache.http.HttpStatus;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
