+++ src/test/java/kitchenpos/table/domain/TableGroupTest.java
@@ -1,6 +1,7 @@
 package kitchenpos.table.domain;
 
 import kitchenpos.table.TableGenerator;
+import kitchenpos.tableGroup.domain.TableGroup;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -20,8 +21,8 @@ class TableGroupTest {
     @DisplayName("단체 지정 생성 시 주문 테이블 수가 1개 이하이면 예외가 발생해야 한다")
     @Test
     void createTableGroupByContainUnderOneOrderTable() {
-        assertThatIllegalArgumentException().isThrownBy(() -> 테이블_그룹_생성(Collections.singletonList(주문_테이블)));
-        assertThatIllegalArgumentException().isThrownBy(() -> 테이블_그룹_생성(Collections.emptyList()));
+        assertThatIllegalArgumentException().isThrownBy(() -> 테이블_그룹_생성(주문_테이블_목록_생성(Collections.singletonList(주문_테이블))));
+        assertThatIllegalArgumentException().isThrownBy(() -> 테이블_그룹_생성(주문_테이블_목록_생성(Collections.emptyList())));
     }
 
     @DisplayName("그룹 테이블 생성 요청 시 이미 단체 지정에 속한 주문 테이블이 포함되어 있으면 예외가 발생해야 한다")
@@ -32,17 +33,17 @@ class TableGroupTest {
         테이블_그룹에_속해있는_주문_테이블.joinGroup(new TableGroup(주문_테이블_목록_생성(Arrays.asList(주문_테이블, 테이블_그룹에_속해있는_주문_테이블))));
 
         // then
-        assertThatIllegalArgumentException().isThrownBy(() -> 테이블_그룹_생성(Arrays.asList(주문_테이블, 테이블_그룹에_속해있는_주문_테이블)));
+        assertThatIllegalArgumentException().isThrownBy(() -> 테이블_그룹_생성(주문_테이블_목록_생성(Arrays.asList(주문_테이블, 테이블_그룹에_속해있는_주문_테이블))));
     }
 
     @DisplayName("정상 상태의 그룹 테이블 생성 요청 시 정상 동작해야 한다")
     @Test
     void createTableGroupTest() {
         // when
-        TableGroup 테이블_그룹 = 테이블_그룹_생성(Arrays.asList(주문_테이블, 주문_테이블));
+        TableGroup 테이블_그룹 = 테이블_그룹_생성(주문_테이블_목록_생성(Arrays.asList(주문_테이블, 주문_테이블)));
 
         // then
-        assertThat(테이블_그룹.getOrderTables().get(0)).isEqualTo(주문_테이블);
+        assertThat(테이블_그룹.getOrderTables().getValue().get(0)).isEqualTo(주문_테이블);
         assertThat(주문_테이블.getTableGroup()).isEqualTo(테이블_그룹);
     }
 
@@ -50,7 +51,7 @@ class TableGroupTest {
     @Test
     void ungroupTest() {
         // given
-        TableGroup 테이블_그룹 = 테이블_그룹_생성(Arrays.asList(주문_테이블, 주문_테이블));
+        TableGroup 테이블_그룹 = 테이블_그룹_생성(주문_테이블_목록_생성(Arrays.asList(주문_테이블, 주문_테이블)));
 
         // when
         테이블_그룹.ungroup();
